backend/requirements.txt
aiofiles==0.8.0
anyio==3.4.0
asgiref==3.4.1
astroid==2.5.1
asyncpg==0.25.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cos-python-sdk-v5==1.9.14
crcmod==1.7
cryptography==36.0.1
databases==0.5.3
Deprecated==1.2.13
dicttoxml==1.7.4
distlib==0.3.4
ecdsa==0.17.0
et-xmlfile==1.1.0
fastapi==0.70.1
fastapi-pagination==0.9.1
filelock==3.4.2
greenlet==1.1.2
h11==0.12.0
idna==3.3
isort==5.7.0
jwt==1.3.1
lazy-object-proxy==1.5.2
loguru==0.5.3
mccabe==0.6.1
openpyxl==3.0.9
packaging==21.3
passlib==1.7.4
Pillow==9.0.0
pipenv==2021.11.23
platformdirs==2.4.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.12.0
pydantic==1.9.0
pylint==2.7.2
pyparsing==3.0.6
python-jose==3.3.0
python-multipart==0.0.5
qrcode==6.1
redis==4.1.0
requests==2.27.1
rsa==4.8
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
unicorn==1.0.3
urllib3==1.26.8
uvicorn==0.16.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
wifi-password==1.1.1
wrapt==1.12.1
XlsxWriter==3.0.2
gunicorn==20.1.0
